refactor(cli): extract manifest install domain
